requirements.txt
customtkinter
mysql-connector-python
python-dotenev